OPIC-34 1st step of refactoring -- objdef.h
