--- pm_serde_example/log4crc
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!DOCTYPE log4c SYSTEM "">
-
-<log4c>
-  <category name="root" priority="debug" appender="stdout"/>
-  <appender name="stdout" type="stream" layout="basic"/>
-  <layout name="basic" type="basic"/>
-</log4c>
