+++ src/Makefile.am
@@ -1,45 +1,48 @@
 # TODO: Can we put Makefile.am in subdir?
 # https://www.gnu.org/software/automake/manual/html_node/Libtool-Convenience-Libraries.html
-AUTOMAKE_OPTIONS = subdir-objects
 
-AM_CPPFLAGS = -I$(top_srcdir)/src
+SUBDIRS = opic/malloc
 
-lib_LTLIBRARIES = libopic.la
-libopic_la_SOURCES = \
-  opic/common/op_log.c \
-  opic/malloc/span.c \
-  opic/malloc/huge_page.c \
-  opic/malloc/op_heap.c \
-  opic/object/op_serializable.c \
-  opic/object/op_trait.c \
-  opic/object/op_object.c \
-  opic/object/op_malloc_manager.c \
-  opic/object/class_map.c \
-  opic/object/ptr_range_map_splay.c \
-  op_collection.c \
-  op_iterator.c \
-  op_linked_list.c \
-  op_list.c
-libopic_la_CFLAGS = @log4c_CFLAGS@ @PTHREAD_CFLAGS@
-
-include_HEADERS = \
-  opic.h \
-  op_collection.h \
-  op_functional.h \
-  op_iterator.h \
-  op_linked_list.h \
-  op_list.h
-
-# TODO some of the internal headers
-# may not be needed
-nobase_include_HEADERS = \
-  opic/common/op_assert.h \
-  opic/common/op_log.h \
-  opic/common/op_macros.h \
-  opic/malloc/op_heap.h \
-  opic/malloc/huge_page.h \
-  opic/malloc/span.h \
-  opic/object/op_serializable.h \
-  opic/object/op_trait.h \
-  opic/object/op_object_def.h \
-  opic/object/op_object.h
+# AUTOMAKE_OPTIONS = subdir-objects
+# 
+# AM_CPPFLAGS = -I$(top_srcdir)/src
+# 
+# lib_LTLIBRARIES = libopic.la
+# libopic_la_SOURCES = \
+#   opic/common/op_log.c \
+#   opic/malloc/span.c \
+#   opic/malloc/huge_page.c \
+#   opic/malloc/op_heap.c \
+#   opic/object/op_serializable.c \
+#   opic/object/op_trait.c \
+#   opic/object/op_object.c \
+#   opic/object/op_malloc_manager.c \
+#   opic/object/class_map.c \
+#   opic/object/ptr_range_map_splay.c \
+#   op_collection.c \
+#   op_iterator.c \
+#   op_linked_list.c \
+#   op_list.c
+# libopic_la_CFLAGS = @log4c_CFLAGS@ @PTHREAD_CFLAGS@
+# 
+# include_HEADERS = \
+#   opic.h \
+#   op_collection.h \
+#   op_functional.h \
+#   op_iterator.h \
+#   op_linked_list.h \
+#   op_list.h
+# 
+# # TODO some of the internal headers
+# # may not be needed
+# nobase_include_HEADERS = \
+#   opic/common/op_assert.h \
+#   opic/common/op_log.h \
+#   opic/common/op_macros.h \
+#   opic/malloc/op_heap.h \
+#   opic/malloc/huge_page.h \
+#   opic/malloc/span.h \
+#   opic/object/op_serializable.h \
+#   opic/object/op_trait.h \
+#   opic/object/op_object_def.h \
+#   opic/object/op_object.h
