--- src/opic.h
@@ -1,11 +1,0 @@
-#ifndef OPIC_H
-#define OPIC_H 1
-
-#include "opic/common/op_macros.h"
-#include "opic/common/op_assert.h"
-#include "opic/common/op_log.h"
-
-#include "opic/object/op_object.h"
-#include "opic/object/op_serializable.h"
-
-#endif /* OPIC_H */
