+++ src/opic/malloc/objdef.c
@@ -1,0 +1,11 @@
+#include "objdef.h"
+#include <stdio.h>
+#include <inttypes.h>
+
+int main()
+{
+  printf("size of USpan: %zu\n", sizeof(UnarySpan));
+  printf("size of USpanQueue: %zu\n", sizeof(UnarySpanQueue));
+  printf("sizeof(RawType): %zu\n", sizeof(RawType));
+  return 0;
+}
